+++ mate/src/main/java/com/kh/mate/erp/controller/ErpContorller.java
@@ -1,8 +1,17 @@
 package com.kh.mate.erp.controller;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.io.PrintWriter;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,12 +22,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.google.gson.JsonObject;
+import com.kh.mate.common.Utils;
 import com.kh.mate.erp.model.service.ErpService;
 import com.kh.mate.erp.model.vo.EMP;
 import com.kh.mate.product.model.vo.Product;
+import com.kh.mate.product.model.vo.ProductMainImages;
 
 
 @Controller
@@ -178,5 +192,173 @@ public class ErpContorller {
 		mav.setViewName("ERP/empList");
 		return mav;
 
+	}
+	
+	//김종완 상품등록
+
+	//상품 등록 시 jsp연결
+	@RequestMapping(value = "/ERP/productEnroll.do",
+					method = RequestMethod.GET)
+	public String productinsert() {
+		
+		return "/ERP/productEnroll";
+	}
+	
+	//상품 등록 submit버튼 클릭시 db저장
+	@RequestMapping(value = "/ERP/productEnroll.do",
+					method = RequestMethod.POST)
+	public String productEnroll(Product product,
+								@RequestParam("upFile") MultipartFile[] upFiles,
+								HttpServletRequest request) throws IllegalStateException, IOException {
+		
+		//Content내 저장 폴더 명 변경
+		String content = product.getContent();
+		log.debug("content = {}", content);
+		String repCont = content.replaceAll("temp", "images");
+		log.debug("repCont = {}", repCont);
+		product.setContent(repCont);
+		
+		//등록 시 업로드한 MainImages를 List로 가져오기
+		List<ProductMainImages> mainImgList = new ArrayList<>();
+		String saveDirectory = request.getServletContext().getRealPath("/resources/upload/mainimages");
+		
+			for(MultipartFile upFile : upFiles) {
+				
+				//파일을 선택하지 않고 전송한 경우
+				if(upFile.isEmpty()) {
+					//섬네일 이미지는 반드시 필요하기 때문에 제출하지 않은 경우 에러넘겨줘야됨.
+				}
+				//파일을 첨부하고 전송한 경우 
+				else {
+					//1.1  파일명(renamdFilename) 생성
+					String renamedFilename = Utils.getRenamedFileName(upFile.getOriginalFilename());
+					
+					//1.2 메모리(RAM)에 저장되어있는 파일 -> 서버 컴퓨터 디렉토리 저장 tranferTo
+					File dest = new File(saveDirectory, renamedFilename);
+					upFile.transferTo(dest);
+					
+					//1.3 ProductMainImages객체 생성
+					ProductMainImages mainImgs = new ProductMainImages();
+					mainImgs.setOriginalFilename(upFile.getOriginalFilename());
+					mainImgs.setRenamedFilename(renamedFilename);
+					mainImgList.add(mainImgs);
+					
+				}
+				
+			}
+		
+		//Product객체에 MainImages객체를 Setting
+		log.debug("mainImgList = {}", mainImgList);	
+		product.setProductMainImages(mainImgList);
+		
+		//Product객체에 ImagesName Setting
+		String tempDir = request.getServletContext().getRealPath("/resources/upload/temp");
+		String imgDir = request.getServletContext().getRealPath("/resources/upload/images");
+		File folder1 = new File(tempDir);
+		File folder2 = new File(imgDir);
+		
+		List<String> productImages = Utils.getOriginFileNameOnTemp(folder1);
+		log.debug("productImages = {}",productImages);
+		product.setProductImagesName(productImages);
+		
+		log.debug("product = {}", product);
+		int result = erpService.productEnroll(product);
+		
+		//product입력 시, file입력 처리 -> DB에 image등록과 동시에 fileDir옮기기  
+		if(result > 0) {
+			Utils.fileCopy(folder1, folder2);
+			Utils.fileDelete(folder1.toString());
+		}else {
+			Utils.fileDelete(folder1.toString());
+		}
+		
+		return "redirect:/";
+	}
+	
+	//뒤로가기 클릭 시 파일 삭제
+	@RequestMapping(value = "/ERP/fileDelMethod.do",
+					method = RequestMethod.GET)
+	public String fileDelete(HttpServletRequest request) {
+		//파일 삭제
+		String tempDir = request.getServletContext().getRealPath("/resources/upload/temp");
+		File folder1 = new File(tempDir);
+		Utils.fileDelete(folder1.toString());
+		
+		return "redirect:/";
+		
+	}
+	
+	@RequestMapping(value = "/ERP/imageFileUpload.do", method = RequestMethod.POST)
+	@ResponseBody
+	public String fileUpload(HttpServletRequest request, HttpServletResponse response,
+							 MultipartHttpServletRequest multiFile) throws Exception {
+		request.setCharacterEncoding("utf-8");
+		JsonObject json = new JsonObject();
+		PrintWriter printWriter = null;
+		OutputStream out = null;
+		MultipartFile file = multiFile.getFile("upload");
+		
+		if(file != null) {
+			if(file.getSize() > 0 ) {
+				if(file.getContentType().toLowerCase().startsWith("image/")) {
+					try {
+						String fileName = Utils.getRenamedFileName(file.getOriginalFilename());
+						byte[] bytes = file.getBytes();
+						String uploadPath = request.getServletContext().getRealPath("/resources/upload/temp");
+						File uploadFile = new File(uploadPath);
+						if(!uploadFile.exists()) {
+							uploadFile.mkdirs();
+						}
+						//String renamedFilename = Utils.getRenamedFileName(fileName);
+						//uploadPath = uploadPath + "/" + fileName;
+						out = new FileOutputStream(new File(uploadPath, fileName));
+						out.write(bytes);
+						
+						printWriter = response.getWriter();
+						response.setContentType("text/html");
+						String fileUrl = request.getContextPath() + "/resources/upload/temp/" + fileName;
+						
+						//json 데이터로 등록
+						json.addProperty("uploaded", 1);
+						json.addProperty("fileName", fileName);
+						json.addProperty("url", fileUrl);
+						
+						printWriter.println(json);
+						
+					} catch(IOException e) {
+						e.printStackTrace();
+					} finally {
+						if(out != null)
+							out.close();
+						if(printWriter != null)
+							printWriter.close();
+					}
+				}
+			}
+		}
+		
+		return null;
+	}
+	
+	//jsp연결
+	@RequestMapping(value = "/ERP/productUpdate.do",
+					method = RequestMethod.GET)
+	public String productUpdate(@RequestParam("productNo") String productNo,
+							  Model model) {
+		
+		Product product = erpService.selectProductOne(productNo);
+		
+		List<ProductMainImages> list = erpService.selectProductMainImages(productNo);
+		model.addAttribute("product", product);
+		model.addAttribute("list", list);
+		
+		return "product/productUpdate";
+	}
+	
+	//수정
+	@RequestMapping(value = "/ERP/productUpdate.do",
+					method = RequestMethod.POST)
+	public String productUpdate() {
+		return null;
 	}
 }
