+++ mate/src/main/java/com/kh/mate/erp/model/dao/ErpDAO.java
@@ -7,6 +7,8 @@ import java.util.Map;
 
 import com.kh.mate.erp.model.vo.EMP;
 import com.kh.mate.product.model.vo.Product;
+import com.kh.mate.product.model.vo.ProductImages;
+import com.kh.mate.product.model.vo.ProductMainImages;
 
 public interface ErpDAO {
 
@@ -19,6 +21,14 @@ public interface ErpDAO {
 	List<EMP> empList();
 
 	Product orderProduct(Map<String, Object> map);
+
+	int productEnroll(Product product);
+
+	int productImageEnroll(ProductImages pigs);
+
+	Product selectProductOne(String productNo);
+
+	List<ProductMainImages> selectProductMainImages(String productNo);
 
 
 }
