+++ mate/src/main/java/com/kh/mate/product/model/vo/ProductImages.java
@@ -16,7 +16,7 @@ import lombok.ToString;
 public class ProductImages {
 	
 	private int productImageNo;
-	private String originalFilename;
+//	private String originalFilename;
 	private String renamedFilename;
 	private int productNo;
 	//private String empId2;
