+++ mate/src/main/resources/mapper/erp/erpProduct/erp-mapper.xml
@@ -54,5 +54,56 @@
 			s.emp_id = #{eId}
 	</select>
 	
+	<select id="selectProductOne" resultType="product">
+ 		select
+ 			*
+ 		from
+ 			product 
+ 		where
+ 			product_no = #{ productNo }
+ 	</select>
+ 	
+ 	<select id="selectProductMainImages" resultType="productMainImages">
+ 		select
+ 			*
+ 		from
+ 			product_main_images
+ 		where
+ 			product_no = #{ productNo }
+ 	</select>
+ 	
+	<insert id="productEnroll">
+ 		insert into 
+ 			product
+ 		values
+ 		(
+ 		seq_product_no.nextval,
+ 		 #{productName},
+ 		 default,
+ 		 #{category},
+ 		 #{content},
+ 		 #{price},
+ 		 #{empId},
+ 		 default
+ 		)
+ 		<selectKey order="AFTER" resultType="_int" keyProperty="productNo">
+			select
+				seq_product_no.currval 			
+ 			from
+ 				dual
+ 		</selectKey>
+ 	</insert>
+ 
+ 	
+ 	<insert id="productImageEnroll">
+ 		insert into
+ 			product_images
+ 		values
+ 		(
+		seq_product_image_no.nextval,
+		#{ renamedFilename },
+		#{ productNo }
+ 		)
+ 	</insert>
 
 </mapper>
