+++ mate/src/main/resources/mapper/product/product-mapper.xml
@@ -37,51 +37,6 @@
  			product_no = #{ productNo }
  	</select>
  	
- 	<insert id="productEnroll">
- 		insert into 
- 			product
- 		values
- 		(
- 		seq_product_no.nextval,
- 		 #{productName},
- 		 default,
- 		 #{category},
- 		 #{content},
- 		 #{price},
- 		 #{empId},
- 		 default
- 		)
- 		<selectKey order="AFTER" resultType="_int" keyProperty="productNo">
-			select
-				seq_product_no.currval 			
- 			from
- 				dual
- 		</selectKey>
- 	</insert>
- 	
- 	
- 	<insert id="mainImegesEnroll">
- 		insert into
- 			PRODUCT_MAIN_IMAGES
- 		values(
- 			seq_product_main_image_no.nextval,
- 			#{ originalFilename },
- 			#{ renamedFilename },
- 			#{ productNo }
- 		)
- 	</insert>
- 	
- 	
- 	<insert id="productImageEnroll">
- 		insert into
- 			product_images
- 		values(
- 			seq_product_image_no.nextval,
- 			#{ renamedFilename },
- 			#{ productNo }
- 		)
- 	</insert>
- 	
  	
  	<select id="searchProductList" resultType="product">
  		select * from product
